# web/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LocationType": {
        "enum": [
          "online",
          "physical",
          "hybrid"
        ],
        "title": "LocationType",
        "type": "string"
      },
...
      "ServiceCreate": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "address": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Address"
              },
              {
                "type": "null"
              }
            ]
          },
          "bookingMode": {
            "enum": [
              "open",
              "scheduled"
            ],
            "title": "Bookingmode",
            "type": "string"
          },
          "cancellationRule": {
            "default": "",
            "maxLength": 1000,
            "title": "Cancellationrule",
            "type": "string"
          },
          "capacity": {
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Capacity",
            "type": "integer"
          },
          "description": {
            "default": "",
            "maxLength": 4000,
            "title": "Description",
            "type": "string"
          },
          "durationMin": {
            "maximum": 1440.0,
            "minimum": 5.0,
            "title": "Durationmin",
            "type": "integer"
          },
          "location": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Location",
            "type": "string"
          },
          "locationType": {
            "enum": [
              "online",
              "physical",
              "hybrid"
            ],
            "title": "Locationtype",
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "priceCents": {
            "minimum": 0.0,
            "title": "Pricecents",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "durationMin",
          "priceCents",
          "capacity",
          "locationType",
          "location",
          "bookingMode"
        ],
        "title": "ServiceCreate",
        "type": "object"
      },
      "ServiceListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ServiceResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ServiceListResponse",
        "type": "object"
      },
...
      "ServiceResponse": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "address": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Address"
              },
              {
                "type": "null"
              }
            ]
          },
          "bookingMode": {
            "enum": [
              "open",
              "scheduled"
            ],
            "title": "Bookingmode",
            "type": "string"
          },
          "cancellationRule": {
            "title": "Cancellationrule",
            "type": "string"
          },
          "capacity": {
            "title": "Capacity",
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "durationMin": {
            "title": "Durationmin",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "locationType": {
            "enum": [
              "online",
              "physical",
              "hybrid"
            ],
            "title": "Locationtype",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "priceCents": {
            "title": "Pricecents",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "durationMin",
          "priceCents",
          "currency",
          "capacity",
          "locationType",
          "location",
          "address",
          "bookingMode",
          "cancellationRule",
          "active",
          "notes",
          "createdAt",
          "updatedAt"
        ],
        "title": "ServiceResponse",
        "type": "object"
      },
...
      "WorkspaceLocationCreate": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "label": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label",
          "address"
        ],
        "title": "WorkspaceLocationCreate",
        "type": "object"
      },
      "WorkspaceLocationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceLocationResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "WorkspaceLocationListResponse",
        "type": "object"
      },
      "WorkspaceLocationPatch": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Address"
              },
              {
                "type": "null"
              }
            ]
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "title": "WorkspaceLocationPatch",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Slotera API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/services": {
      "get": {
        "operationId": "listServices",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 160,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Active"
            }
          },
          {
            "in": "query",
            "name": "locationType",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LocationType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Locationtype"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Services",
        "tags": [
          "services"
        ]
      },
      "post": {
        "operationId": "createService",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Service",
        "tags": [
          "services"
        ]
      }
    },
    "/services/{service_id}": {
      "delete": {
        "operationId": "deleteService",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Service Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Service",
        "tags": [
          "services"
        ]
      },
      "get": {
        "operationId": "getService",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Service Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Service",
        "tags": [
          "services"
        ]
      },
      "patch": {
        "operationId": "updateService",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Service Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServicePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Service",
        "tags": [
          "services"
        ]
      }
    },
...
    "/settings/business": {
      "get": {
        "operationId": "getBusinessSettings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Business Settings",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "updateBusinessSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessSettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Business Settings",
        "tags": [
          "settings"
        ]
      }
    },
    "/settings/locations": {
      "get": {
        "operationId": "listWorkspaceLocations",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceLocationListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Workspace Locations",
        "tags": [
          "settings"
        ]
      },
      "post": {
        "operationId": "createWorkspaceLocation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceLocationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceLocationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Workspace Location",
        "tags": [
          "settings"
        ]
      }
    },
    "/settings/locations/{location_id}": {
      "delete": {
        "operationId": "deleteWorkspaceLocation",
        "parameters": [
          {
            "in": "path",
            "name": "location_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Location Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Workspace Location",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "updateWorkspaceLocation",
        "parameters": [
          {
            "in": "path",
            "name": "location_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Location Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceLocationPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceLocationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Workspace Location",
        "tags": [
          "settings"
        ]
      }
    },
    "/settings/payments": {
      "get": {
        "operationId": "getPaymentSettings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Payment Settings",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "updatePaymentSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentSettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Payment Settings",
        "tags": [
          "settings"
        ]
      }
    }
  }
}